:christmas_tree: Added support for Lumberyard Version 1.12
Added new logo for Copal that shows up in project configurator.
Added PhysicsIncludes.h which includes LmbrCentral/Physics/* files and sets up a typedef to allow switching between LY 1.11 and 1.12. Default is set to LY 1.12.
Changed to use CryPhysicsComponentRequestBus for LY 1.12
Added a typedef  to switch back to the old PhysicsComponentRequest bus for LY 1.11 usage.